The minimal change to make it work with version 29
New files added in the manifest, commented the version <4 entry.
I find no quick way to keep the version<4 compatibility because sadly
the appversion flag is not meant for 3 seperate versions. You may want
to drop that compatibility?

New files created for use in version 29 :
In the xul : creation of a new bar in the bottom box instead of using
the addon bar.
In the js : just commented the small part about organizing the addon
bar.