Always hide blank tab; hide on starting FF; set value by attribute.
Sometimes a blank tab had values (maybe because of Super Start) and so it
wouldn't hide using "All Protocols", so "about:blank" is explicitly hidden.

If Firefox is started with a blank tab, the style would still be visible,
so hide ESB's status bar on init.

A strange problem occurred when customizing from a second blank tab or
from Super Start (which creates a new tab for customizing) - values would
not update.  Turns out `.value = string` would not work, but
`.setAttribute("value", string)` would, so to be safe the labels are now
`set` properties, using `setAttribute`.